// src/app/models/Emotion.ts
export enum Emotion {
  Angry = 'angry',
  Base = 'base',
  Happy = 'happy',
  Sad = 'sad',
  Thinker = 'thinker',
  WatcherBottom = 'watcher_bottom',
  WatcherLeft = 'watcher_left',
  WatcherRight = 'watcher_right',
  WatcherTop = 'watcher_top',
  WatcherBottomLeft = 'watcher_bottom_left',
  WatcherBottomRight = 'watcher_bottom_right',
  WatcherTopLeft = 'watcher_top_left',
  WatcherTopRight = 'watcher_top_right',
  Writing1 = 'writing_1',
  Writing2 = 'writing_2',
  WritingReady = 'writing_ready',
}